made go to dashboard log in and take you to dashboard
