+++ utils/constants.ts
@@ -7,6 +7,7 @@ import codeexamplesHoverCSNAV from "@/public/sidenav/card-demo-sidenav-codeexamp
 import codeexamplesNoHoverCSNAV from "@/public/sidenav/card-demo-sidenav-codeexamples.svg";
 import curlyBrackets from "@/public/sidenav/curly-brackets.svg";
 import { AIModelInterface, UserDataType } from "./typescriptTypesInterfaceIndustry";
+import { Persona } from "@/utils/typescriptTypesInterfaceLogin";
 
 export const ALERT_TYPES = {
     SUCCESS: "success",
@@ -101,16 +102,55 @@ export const NAV_ELEMENTS_VARIANT = {
     },
 };
 
+export const STARTER_PERSONAS: Persona[] = [
+    {
+        personaname: "Christine",
+        personatier: PERSONA_TIER_STANARD,
+        personaimage: "/personas/persona3.png",
+        personaemail: "user@launchmail.io",
+        personarole: PERSONA_ROLE_USER,
+    },
+    {
+        personaname: "Angela",
+        personatier: PERSONA_TIER_PLATINUM,
+        personaimage: "/personas/persona6.jpg",
+        personaemail: "angela@launchmail.io",
+        personarole: PERSONA_ROLE_USER,
+    },
+    {
+        personaname: "Alysha",
+        personatier: PERSONA_TIER_STANARD,
+        personaimage: "personas/beta.png",
+        personaemail: "alysha@launchmail.io",
+        personarole: PERSONA_ROLE_BETA,
+    },
+    {
+        personaname: "Jenn",
+        personatier: PERSONA_TIER_STANARD,
+        personaimage: "personas/woman.png",
+        personaemail: "jenn@launchmail.io",
+        personarole: PERSONA_ROLE_DEVELOPER,
+    },
+    {
+        personaname: "Cody",
+        personatier: PERSONA_TIER_STANARD,
+        personaimage: "personas/standard.jpg",
+        personaemail: "cody@launchmail.io",
+        personarole: PERSONA_ROLE_USER,
+    },
+];
+
+
 export const INITIAL_USER_SIGNUP_DATA: UserDataType = {
-  email: "newUser@launchmail.io",
+  email: STARTER_PERSONAS[0].personaemail,
   password: "defaultPassword",
-  firstName: "",
-  lastName: "",
-  dob: "",
-  ssn: "",
-  phone: "",
-  address: "",
-  apt: "",
-  zip: "",
-  selectedServices: [	"Checking Accounts"],
+  firstName: STARTER_PERSONAS[0].personaname,
+  lastName: "Wilson",
+  dob: "2/28/1998",
+  ssn: "***-**-****",
+  phone: "220-415-9634",
+  address: "390 Fort St",
+  apt: "245",
+  zip: "94572",
+  selectedServices: ["Checking Accounts"],
 }
